Decompose Large and Sparse Matrix

Computes Eigenvalues of the similarity matrix. As input, the contingency matrix M should be given. See for a elaborate description van Dam, et al, (2021)

decomp_symmetric(matrix, max_eigenvalues, decomp = "svd")

Arguments

matrix
Incidence matrix (e.g. species - location), which can be interpreted as the bi-adjacency matrix of a bipartite network.

max_eigenvalues
Max. number of eigenvalues to compute. Default is 25.

decomp
The decomposition to use: svd (default) or svds. The later only computes the k leading singular values and vectors of a rectangular matrix

Details

If both dimensions of data matrix are greater than max_eigenvalues, then the number of computed eigenvalues is restricted to max_eigenvalues; otherwise, the shortest dimension is chosen.

svds uses the rARPACK implementation of the singular value decomposition for efficient approximation of singular decomposition for large sparse matrices.

References

van Dam, et al. (2021), Correspondence analysis, spectral clustering and graph embedding: applications to ecology and economic complexity; *name of journal*; DOI: <doi>.