Print SCCA tree

Prints the output of compute_scca to screen in a human-readable format.

scca_print(scca, ...)

Arguments

scca
An output of scca_compute to screen

...
The attributes to be printed: 'depth', 'k', 'node_type' and/or 'n_labs'

Details

The nodes in the analysis tree (clusters) are numbered depth-first and pre-order. depth is the level in the tree. n_labs is the number of of labels in the (sub-)cluster. child is the numbering within its siblings. And k is the number of relevant Eigenvalues. Every attribute can be omitted. If no attribute is given, only the tree hierarchy will be printed.

Examples

  if (FALSE) {
     s <- scca_compute(t(carnivora))
     scca_print(scca = s, 'k', 'n_labs')
}